Simplified organization of Rayon-based parallel iterators in GraphMap.
